refactor: suspense 적용 및 컴포넌트 이름 변경
