+++ src/app/map/page.module.scss
@@ -34,14 +34,10 @@
     z-index: 1;
     overflow: hidden;
 
-    gap: 24px;
-
     padding: 24px 24px 152px;
 
     background-color: $white;
     border-radius: 24px 24px 0;
-
-    @include flexColumnStart;
   }
 
   .urlButtonContainer {
